+++ CHANGELOG.md
@@ -11,6 +11,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Added option to `ignore_color_table` when using gdal imread backend.
 
 
+### Changes
+
+* Fixes for `intensity_normalize` 
+* Allow user to specify nodata, transform, and crs when writing a geotiff
+
+
 ### Changed
 * Consolidated environment variables in a new `kwiamge._internal` module.